+++ example-app/src/js/example.js
@@ -40,6 +40,17 @@ document.addEventListener('DOMContentLoaded', () => {
     const result = await CapacitorCalendar.requestFullRemindersAccess();
     console.log('#requestFullRemindersAccess', result);
   });
+
+  document.querySelector('#update-reminders-list').addEventListener('click', async () => {
+    const result = await CapacitorCalendar.updateRemindersList({
+      color: 'indigo',
+      id: getRemindersListIdInput().value,
+      title: 'Updated Groceries list',
+    });
+
+    getRemindersListIdInput().value = result.id;
+    console.log('#updateRemindersList', result);
+  });
 });
 
 function getRemindersListIdInput() {