+++ ios/Plugin/Models/Inputs/UpdateRemindersListInput.swift
@@ -1,0 +1,55 @@
+import Capacitor
+
+struct UpdateRemindersListInput {
+    private let color: CGColor?
+    private let commit: Bool
+    private let id: String
+    private let title: String?
+
+    init(call: CAPPluginCall) throws {
+        guard let id = call.getString("id") else {
+            throw PluginError.idMissing
+        }
+        self.id = id
+        self.title = call.getString("title")
+        self.color = UpdateRemindersListInput.getListColorFromCall(call)
+        self.commit = call.getBool("commit", true)
+    }
+
+    func getColor() -> CGColor? {
+        return color
+    }
+
+    func getCommit() -> Bool {
+        return commit
+    }
+
+    func getId() -> String {
+        return id
+    }
+
+    func getTitle() -> String? {
+        return title
+    }
+
+    private static func getListColorFromCall(_ call: CAPPluginCall) -> CGColor? {
+        guard let colorName = call.getString("color") else {
+            return nil
+        }
+
+        switch colorName.lowercased() {
+        case "blue":    return UIColor.systemBlue.cgColor
+        case "brown":   return UIColor.systemBrown.cgColor
+        case "gray":    return UIColor.systemGray.cgColor
+        case "green":   return UIColor.systemGreen.cgColor
+        case "indigo":  return UIColor.systemIndigo.cgColor
+        case "orange":  return UIColor.systemOrange.cgColor
+        case "pink":    return UIColor.systemPink.cgColor
+        case "purple":  return UIColor.systemPurple.cgColor
+        case "red":     return UIColor.systemRed.cgColor
+        case "teal":    return UIColor.systemTeal.cgColor
+        case "yellow":  return UIColor.systemYellow.cgColor
+        default:        return nil
+        }
+    }
+}